feat: create useInfinityScroll hook
