+++ src/pages/archive/index.tsx
@@ -1,10 +1,9 @@
 import API from '@/api/API';
 import { LinkItemWithProfileList } from '@/components/LinkItem';
-import useIntersectionObserver from '@/hooks/useIntersectionObserver';
+import useInfinityScroll from '@/hooks/useInfinityScroll';
 import { useAppDispatch } from '@/store';
 import { routerSlice } from '@/store/slices/routerSlice';
-import { useInfiniteQuery } from '@tanstack/react-query';
-import React, { useCallback, useEffect, useRef } from 'react';
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
 
 const Explore = () => {
@@ -16,13 +15,10 @@ const Explore = () => {
 
   const isUser = false;
   const fetchLinksFn = isUser ? API.getAuthLinksArchive : API.getLinksArchive;
-  const target = useRef(null);
 
-  const { data, fetchNextPage, hasNextPage, isFetchingNextPage } = useInfiniteQuery({
+  const { pages, target, isFetchingNextPage } = useInfinityScroll({
+    fetchFn: fetchLinksFn,
     queryKey: ['archive'],
-    queryFn: ({ pageParam }) => {
-      return fetchLinksFn(pageParam || '');
-    },
     getNextPageParam: (lastPage_) => {
       const lastPage = lastPage_.data?.linkArchive;
       const lastItem = lastPage[lastPage.length - 1].urlId;
@@ -30,28 +26,11 @@ const Explore = () => {
 
       return hasNext ? lastItem : undefined;
     },
-    retry: 1,
   });
-
-  const pages = data?.pages || [];
-
-  const handleIntersection = useCallback(
-    (entries) => {
-      entries.forEach((entry) => {
-        if (entry.isIntersecting && !isFetchingNextPage) {
-          hasNextPage && fetchNextPage();
-        }
-      });
-    },
-    [fetchNextPage, hasNextPage, isFetchingNextPage]
-  );
-  useIntersectionObserver({ target, onIntersection: handleIntersection });
 
   return (
     <Wrapper>
-      {pages.map(({ data: { linkArchive } }) => (
-        <LinkItemWithProfileList linkList={linkArchive} key='' />
-      ))}
+      <LinkItemWithProfileList data={pages} />
       {isFetchingNextPage && <div>로딩중...</div>}
       <div ref={target} />
     </Wrapper>
